mi64: stop viewing uint64 arrays through uint32 pointers
mi64_is_div_by_scalar32() and its _x4/_x8/32p siblings declared their input as
const uint32 x[] and doubled the caller-supplied (uint64) length internally, so
every call site handed them a uint64 array through a (uint32 *) cast. The
comment above the first of them described the resulting incompatible-pointer
warnings as "expected".

That is a strict-aliasing violation, and clang 14 on aarch64 acts on it: with
-flto the call is inlined, TBAA concludes the uint32 reads cannot alias the
uint64 object, and the function returns FALSE for inputs that are plainly
divisible. mi64_div_y32() then trips its own cross-check:

  Assertion 'mi64_is_div_by_scalar32((const uint32 *)x, y, len)' failed:
  Results of mi64_div_y32 and mi64_is_div_by_scalar32 differ!

Instrumenting the assertion showed the disagreements are on trivial inputs -
255/3, 255/5, 763/7, 2541/11, 2795/13 - all genuinely divisible, i.e.
mi64_div_y32 was right and the cross-check wrong. Only the check is affected;
the division results callers use were always correct.

Take const uint64 x[] instead and read each half via

  static uint32 mi64_word32(const uint64 x[], uint32 i)
  { return (uint32)(x[i>>1] >> ((i&1)<<5)); }

which is by value, so there is no aliasing violation - and, because the shift
addresses the value rather than the storage, no dependence on host byte order.
The old word indexing assumed little-endian; this is one less thing for a
big-endian port to trip over. mi64_is_div_by_scalar32p_x8() is left alone: it
takes eight separate arrays and has no callers.

Verified on aarch64 (Ubuntu clang 14.0.0, -O3 -flto, under qemu), same branch
and container throughout:

  unpacked, no fix            5 disagreements, abort
  no fix + -fno-strict-alias  0            (workaround, confirms the mechanism)
  no fix + -fno-lto           0            (workaround, confirms the mechanism)
  this fix, no workarounds    0, exit 0, factoring self-tests pass

On x86-64: Mlucas builds clean in nosimd and avx2, and Mfactor's self-tests -
which call the _x4 and _x8 variants directly - pass.

+++ src/mi64.h
@@ -317,9 +317,9 @@ DEV uint64	mi64_div_by_scalar64_u4	(uint64 x[], uint64 a, uint32 lenX, uint64 q[
 	int		mi64_div_binary			(const uint64 x[], const uint64 y[], uint32 lenX, uint32 lenY, uint64 q[], uint32*lenQ, uint64 r[]);
 
 /* Fast is-divisible-by-scalar, div-by-scalar y]: */
-DEV int		mi64_is_div_by_scalar32 (const uint32 x[], uint32 a, uint32 len);
-DEV uint32	mi64_is_div_by_scalar32_x4(const uint32 x[], uint32 tryq0, uint32 tryq1, uint32 tryq2, uint32 tryq3, uint32 len);
-DEV uint32	mi64_is_div_by_scalar32_x8(const uint32 x[], uint32 tryq0, uint32 tryq1, uint32 tryq2, uint32 tryq3, uint32 tryq4, uint32 tryq5, uint32 tryq6, uint32 tryq7, uint32 len);
+DEV int		mi64_is_div_by_scalar32 (const uint64 x[], uint32 a, uint32 len);
+DEV uint32	mi64_is_div_by_scalar32_x4(const uint64 x[], uint32 tryq0, uint32 tryq1, uint32 tryq2, uint32 tryq3, uint32 len);
+DEV uint32	mi64_is_div_by_scalar32_x8(const uint64 x[], uint32 tryq0, uint32 tryq1, uint32 tryq2, uint32 tryq3, uint32 tryq4, uint32 tryq5, uint32 tryq6, uint32 tryq7, uint32 len);
 DEV int		mi64_is_div_by_scalar64	(const uint64 x[], uint64 a, uint32 len);
 DEV int		mi64_is_div_by_scalar64_x4(const uint64 x[], uint64 q0, uint64 q1, uint64 q2, uint64 q3, uint32 len);
 DEV int		mi64_is_div_by_scalar64_u2	(const uint64 x[], uint64 a, uint32 len);	// 2-way interleaved-|| loop
